Support 16 bits tcpControlFlags in the Ipfix Aggregator
The new Ipfix RFCs (7011 and 7012, see
http://www.iana.org/assignments/ipfix/ipfix.xhtml) have changed
the size of the tcpControlFlags from 1 byte to 2 bytes.
Some part of the code, like the ipfixlolib, already support it,
but the Ipfix Aggregator asserts when the field is not 1 byte long.
The RFC and IANA page specifically say that in case an exporter does
not support the upper flags, then 1 byte should be exported. So, in
Vermont, we have to support both cases in ingress. But internally, in
the aggregator, store everything in 2 bytes format. So if there is no
aggregator and an exporter sends 1 byte field, we export 1 byte. But
if there is an aggregator, convert to 2 bytes.

+++ src/modules/ipfix/IpfixCsExporter.hpp
@@ -120,7 +120,7 @@ class IpfixCsExporter : public Module, public Source<NullEmitable*>, public Ipfi
 			uint16_t destination_transport_port;    // encode udp/tcp ports here
 			uint8_t  icmp_type_ipv4;
 			uint8_t  icmp_code_ipv4;
-			uint8_t  tcp_control_bits;
+			uint16_t  tcp_control_bits;
 			uint64_t flow_start_milliseconds;       // encode flowStart(Micro|Nano|)Seconds here
 			uint64_t flow_end_milliseconds;         // encode flowEnd(Micro|Nano|)Seconds here
 			uint64_t octet_total_count;
@@ -128,7 +128,7 @@ class IpfixCsExporter : public Module, public Source<NullEmitable*>, public Ipfi
 			uint8_t  biflow_direction;
 			uint64_t rev_octet_total_count;
 			uint64_t rev_packet_total_count;
-			uint8_t  rev_tcp_control_bits;
+			uint16_t  rev_tcp_control_bits;
 		} DISABLE_ALIGNMENT;
 
 		list<Ipfix_basic_flow*> chunkList;
